Update Edit button in docs
Since docs moved to another repo, the "edit this page" button started to fail. I validated this change by running "yarn docusaurus start" and pressing on some "edit this page" buttons on pages. This is all I expect to change with this PR, but I am not familiar with Docusaurus to prove it for certain. :)

Signed-off-by: Davide Cafaro <davidecafaro@microsoft.com>